Add solver.
Currently the output for the Solver is horribly broken. It does successfully find the words, but seems to take the end coordinate of the word rather than the start coordinate, and every coordinate is printed as one less than what it should be (meaning that the origin is -1, -1\!). I'll fix it when I have time.