Provide initial command-line arguments
This commit provides command-line argument processing with a system that should be relatively easy to add to. Additionally, it provides two command-line arguments, both with a full name and short alias. One determines if found words should be coloured on the grid, and one provides a custom word count. As part of the implementation of the former, the fix for the solved board not featuring coloured words was discovered and implemented. I had simply failed to call the method to do so.